fix(lint): clear every lib clippy error behind the CI lint job (--all-features, -D warnings)
GitHub CI's lint job has never been green: `cargo clippy --all-targets
--all-features -- -D warnings` failed inside the library, so no test, bench,
or example target was ever linted. This clears the library layer: 19 lints
across 10 files, each fixed per the lint's advice with no behaviour change
and no blanket allows.

- future_not_send: `Send + Sync` bounds on `Http1Listener::run_with`'s
  handler/spawner and `T: Serialize + ?Sized + Sync` on the remote frame
  writers (the lint tolerates only top-level `Param: Send` obligations).
- manual_is_variant_and, let_and_return, unnecessary_map_or, needless_borrow,
  needless_continue, useless_let_if_seq, unnested_or_patterns (h2 listener x2,
  web router), redundant_closure (web response x2), manual_async_fn (extract,
  multipart), unnecessary_fallible_conversions (metadata.rs: i32 narrowing
  then infallible widening to c_long, which also compiles on 32-bit unix
  where the lint's literal suggestion would not).
- rustfmt applied to tests/proof_status_snapshot_contract.rs (the Check job's
  other failure; whitespace-only apart from one redundant closure brace).

Evidence: `cargo clippy -p asupersync --lib --all-features -- -D warnings`
exit 0 on rch (hz3) against the working tree that includes these files.
No-claim: test/bench/example targets are the next layer and were not linted
by this run.

+++ src/net/atp/transport_common/metadata.rs
@@ -2417,9 +2417,12 @@ fn set_special_file_mtime_no_open(path: &Path, seconds: i64, nanos: u32) -> Resu
 
     let seconds = libc::time_t::try_from(seconds)
         .map_err(|_| "mtime seconds out of platform range".to_string())?;
-    let nanos = libc::c_long::try_from(nanos % 1_000_000_000)
+    // `nanos % 1_000_000_000` always fits in `i32`, and `c_long` is at least
+    // 32 bits wide on every unix target, so widening it is infallible even
+    // where `c_long` is `i32` (32-bit targets) and `u32 -> c_long` is not.
+    let nanos = i32::try_from(nanos % 1_000_000_000)
         .map_err(|_| "mtime nanoseconds out of platform range".to_string())?;
-    let mtime = TimeSpec::new(seconds, nanos);
+    let mtime = TimeSpec::new(seconds, libc::c_long::from(nanos));
     utimensat(
         AT_FDCWD,
         path,
